+++ pydynpd/variable.py
@@ -17,8 +17,3 @@ class adjustable_lag_indep:
         self.min_lag = min_lag
         self.max_lag = max_lag
 
-# class indep_var(rhs_var):
-#     pass
-#
-# class iv_var (rhs_var):
-#     pass
