+++ pydynpd/specification_tests.py
@@ -4,7 +4,7 @@ import numpy as np
 from scipy import stats
 
 from pydynpd.info import hansen_test_info, AR_test_info
-
+from pydynpd.common_functions import lag
 
 def hansen_overid(W2_inv, N, zs, num_instru, num_indep):
     hansen_test = np.linalg.multi_dot([zs.transpose(), W2_inv, zs]) * (1.0 / N)
@@ -16,99 +16,74 @@ def hansen_overid(W2_inv, N, zs, num_instru, num_indep):
 
     return (hansen_test)
 
+def AR_get_diff_XR(model, beta, ori_residual,r0_height):
+    N=model.N
 
-def AR_test(model, step, m):  # N, H, M, z_list, XZ_W,vcov, residual,residual_t, Cx_list, level, m):
+    ori_x = model.final_xy_tables['Cx']
+    x0_height=int(ori_x.shape[0]/N)
 
+    if model.options.transformation=='fod':
+        diff_y = model.final_xy_tables['Diff_y']
+        diff_x = model.final_xy_tables['Diff_x']
+
+        diff_r = diff_y - diff_x @ beta
+    elif model.options.level:
+        diff_width = model.z_information.diff_width
+        num_col=ori_x.shape[1]
+        diff_r = np.empty((diff_width * N, 1), dtype=np.float64)
+        diff_x = np.empty((diff_width * N, num_col), dtype=np.float64)
+        for i in range(N):
+            r_i = ori_residual[(i * r0_height):(i * r0_height + r0_height), :]
+            diff_r[(i * diff_width):(i * diff_width + diff_width), 0:1] = r_i[0:diff_width, 0:1]
+            
+            x_i=ori_x[(i * r0_height):(i * r0_height + r0_height), :]
+            diff_x[(i * diff_width):(i * diff_width + diff_width), :] = x_i[0:diff_width, :]
+    else:
+        diff_x=ori_x
+        diff_r=ori_residual
+
+
+    return (diff_x, diff_r)
+
+def AR_test(model, step, m):
     N = model.N
     z_list = model.z_list
-    Cx = model.final_xy_tables['Cx']
-    Cx_list = Cx.dat
-
-    step1 = model.step_results[0]
-    step2 = model.step_results[1]
-    if step == 2:
-        current_step = step2
-    elif step == 1:
-        current_step = step1
-    else:
-        current_step = model.step_results[step - 1]
-
-    M = current_step.M
-    XZ_W = current_step._XZ_W
-    M_XZ_W = current_step._M_XZ_W  # ??????????????????????????????
-    vcov = current_step.vcov
-    residual = current_step.residual
-    residual_t = current_step._residual_t
-
-    r_height = int(residual.shape[0] / N)
-
-    x_height = Cx.unit_height
     z_height = int(z_list.shape[0] / N)
-    x_width = Cx.width
 
-    diff_width = model.z_information.diff_width
+    current_step = model.step_results[step - 1]
 
-    if model.options.level:
+    ori_residual = current_step.residual
+    r0_height = int(ori_residual.shape[0] / N)
 
-        # r_list=[]
-        # r_t_list=[]
-        # x_list=[]
-        r_list = np.empty((diff_width * N, 1), dtype=np.float64)
-        r_t_list = np.empty((1, diff_width * N), dtype=np.float64)
-        x_list = np.empty((diff_width * N, x_width), dtype=np.float64)
+    M_XZ_W = current_step._M_XZ_W
+    vcov = current_step.vcov
 
-        for i in range(N):
-            r = residual[(i * r_height):(i * r_height + r_height), :]
[…]
