+++ pydynpd/instruments.py
@@ -8,20 +8,21 @@ class instruments(object):
     def __init__(self, variables: dict, gmm_tables: dict, df_information: df_info, options: options_info):
         level = options.level
         collapse = options.collapse
+        transformation = options.transformation
         self.z_information, self.gmm_diff_info, self.iv_diff_info, self.gmm_level_info = self.calculate_z_dimension(
-            variables, df_information, level, collapse)
+            variables, df_information, level, transformation, collapse)
 
         if level:
-            self.z_table = self.build_z_level(variables, gmm_tables, df_information, collapse)
+            self.z_table = self.build_z_level(variables, gmm_tables, df_information, transformation, collapse)
         else:
-            self.z_table = self.build_z_diff(variables, gmm_tables, df_information, False, collapse)
+            self.z_table = self.build_z_diff(variables, gmm_tables, df_information, False, transformation, collapse)
 
-    def build_z_level(self, variables: dict, gmm_tables: dict, info: df_info, collapse=False):
-        last_level_index = info.last_index
+    def build_z_level(self, variables: dict, gmm_tables: dict, info: df_info, transformation, collapse=False):
+        last_level_index = info.last_level_index
         first_level_index = info.first_level_index
 
         z_table = self.build_z_diff(
-            variables, gmm_tables, info, True, collapse)
+            variables, gmm_tables, info, True, transformation, collapse)
 
         level_width = self.z_information.level_width
         level_height = self.z_information.level_height
@@ -32,17 +33,16 @@ class instruments(object):
 
         Lgmm_vars = variables['Lgmm']
         iv_vars = variables['iv']
-        Lgmm = gmm_tables['Lgmm']
-        iv = gmm_tables['iv']
-        Lgmm_dat = Lgmm.dat
-        iv_dat = iv.dat
-        Lgmm_iv_height = Lgmm.unit_height
+        Lgmm_dat= gmm_tables['Lgmm']
+        iv_dat = gmm_tables['iv']
+
+        Lgmm_iv_height = int(Lgmm_dat.shape[0]/info.N)
 
         start_row = diff_height  # z_table[0].shape[0]-height
         start_col = diff_width  # z_table[0].shape[1]-width
 
         for i in range(info.N):
-            z = z_table[(i * height):(i * height + height),:]
+            z = z_table[(i * height):(i * height + height), :]
             # z[start_row-1,start_col:(start_col+self.level_width)]=1
             z[height - 1, start_col:width] = 1
             array_Lgmm = Lgmm_dat[(i * Lgmm_iv_height):(i * Lgmm_iv_height + Lgmm_iv_height), :]
@@ -100,13 +100,14 @@ class instruments(object):
 
         num_Lgmm_instr = start_row  # number of gmm instruments in diff eq
 
-        return ((num_Lgmm_instr, t_info))
+        return (num_Lgmm_instr, t_info)
 
-    def calculate_z_dimension(self, variables: dict, info: df_info, level, collapse=False):
+    def calculate_z_dimension(self, variables: dict, info: df_info, level, transformation, collapse=False):
         Lgmm_vars = variables['Lgmm']
 
-        diff_width = info.last_index - info.first_diff_index + 1
-        level_width = info.last_index - info.first_level_index + 1
+        diff_width = info.last_diff_index - info.first_diff_index + 1
+
+        level_width = info.last_level_index - info.first_level_index + 1
 
         level_height = 0
         num_Lgmm_instr = 0
@@ -116,7 +117,7 @@ class instruments(object):
             level_height = num_Lgmm_instr + 1
 
[…]
