+++ pydynpd/command.py
@@ -223,7 +223,7 @@ class command(object):
                 exit()
 
     def parse_options(self, part_3):
-        list_options =[s.lower()   for s in  part_3.split()]
+        list_options = [s.lower() for s in part_3.split()]
 
         options = self.options
 
@@ -236,10 +236,10 @@ class command(object):
                 options.steps = 1000
             elif option == 'nolevel':
                 options.level = False
-            elif option== 'hqic':
-                options.mmsc='hqic'
-            elif option== 'fod':
-                options.transformation='fod'
+            elif option == 'hqic':
+                options.mmsc = 'hqic'
+            elif option == 'fod':
+                options.transformation = 'fod'
             elif option == 'timedumm':
                 options.timedumm = True
             elif option == 'collapse':
