+++ pydynpd/dynamic_panel_model.py
@@ -3,11 +3,13 @@ import math
 import numpy as np
 import pandas as pd
 import scipy
+
 from pydynpd.common_functions import get_first_diff_table, get_fod_table
 from pydynpd.info import df_info, options_info
 from pydynpd.instruments import instruments
 from pydynpd.panel_data import panel_data
 from pydynpd.variable import regular_variable
+
 
 class dynamic_panel_model(object):
     def __init__(self, pdata: panel_data, variables: dict, options: options_info, command_str: str, part_2: str,
@@ -22,21 +24,19 @@ class dynamic_panel_model(object):
         self.command_str = command_str + '|' + part_2
         if part_3 != '':
             self.command_str += '|' + part_3
-        
+
         self.get_info(variables, self.N, self.T)
 
-        
         self.step_results = []
 
         if options.timedumm:
-            if options.transformation=='fd':
+            if options.transformation == 'fd':
                 self.update_time_dummies(self.df_information.first_diff_index, self.df_information.last_diff_index)
             else:
                 self.update_time_dummies(self.df_information.first_fod_index, self.df_information.last_fod_index)
         self.prepare_data()
 
     def update_time_dummies(self, first_diff_index, last_index):
-
 
         for var_name in self.pdata.col_timedumm[first_diff_index:(last_index + 1)]:
             new_var = regular_variable(var_name, 0)
@@ -69,25 +69,23 @@ class dynamic_panel_model(object):
             if var.min_lag > max_Lgmm_minlag:
                 max_Lgmm_minlag = var.min_lag
 
-        last_level_index=T-1
-        last_diff_index=T-1
+        last_level_index = T - 1
+        last_diff_index = T - 1
 
         first_level_index = max(max_lag, max_Lgmm_minlag)
 
-        if self.options.level and self.options.transformation=='fod':
-            first_diff_index =  first_level_index
+        if self.options.level and self.options.transformation == 'fod':
+            first_diff_index = first_level_index
         else:
-            first_diff_index = first_level_index +1   #    max(max_lag + 1, max_Dgmm_minlag)
+            first_diff_index = first_level_index + 1  # max(max_lag + 1, max_Dgmm_minlag)
 
-
-
-        if first_diff_index + 2 > last_diff_index :  # to do: change 3 to something rated to AR(p)
+        if first_diff_index + 2 > last_diff_index:  # to do: change 3 to something rated to AR(p)
             raise Exception("Not enough periods to run the model")
 
         self.df_information = df_info(N=self.N, T=self.T, ids=self.pdata.ids, max_lag=max_lag,
-                                      first_diff_index=first_diff_index, last_diff_index=last_diff_index, first_level_index=first_level_index,
+                                      first_diff_index=first_diff_index, last_diff_index=last_diff_index,
+                                      first_level_index=first_level_index,
                                       last_level_index=last_level_index)
-        
 
     def get_gmm_table_dict(self, variables, level):
         gmm_dict = {}
@@ -119,7 +117,7 @@ class dynamic_panel_model(object):
         ori_Diff_y_table = get_first_diff_table(ori_y_table, self.N)
         ori_Diff_x_table = get_first_diff_table(ori_x_table, self.N)
 
-        if self.options.transformation=='fd':
+        if self.options.transformation == 'fd':
[…]
