+++ vignettes/paper.md
@@ -62,7 +62,7 @@ reason they are not fast), our code uses components common to any
 programming language, making it easy to translate to R or Julia.
 
 <p align = "center">
-<img src = "./../Benchmark/images/Test_1.svg">
+<img src = "https://raw.githubusercontent.com/dazhwu/pydynpd/main/Benchmark/images/Test_1.svg">
 </p>
 <p align = "center">
 Fig.1 - Performance Comparision 
