+++ vignettes/ref.bib
@@ -66,7 +66,7 @@ doi = {10.3390/econometrics8030030}
   pages={e26081},
   year={2021},
   publisher={JMIR Publications Inc., Toronto, Canada},
-  doi={doi: 10.2196/26081}
+  doi={10.2196/26081}
 }
 
 @article{roodman2009xtabond2,
