+++ pydynpd/sandbox/pydynpd/pydynpd/cpp/Common_Functions.cpp
@@ -1,265 +1,0 @@
-//
-// Created by Tiger on 5/4/2022.
-//
-
-#include "Common_Functions.h"
-#define EIGEN_INITIALIZE_MATRICES_BY_NAN
-
-using namespace std;
-using namespace Eigen;
-
-vector<bool> row_has_nan(const RowMatrixXd &x)
-{
-    int num_rows=x.rows();
-    vector<bool> tbr(num_rows);
-    for (int i=0; i<num_rows; ++i){
-
-        tbr[i]=!((x.row(i).array()==x.row(i).array()).all());
-        
-    }
-   return tbr;
-}
-
-RowMatrixXd common_inv(Ref <RowMatrixXd> ori) {
-
-    Eigen::FullPivLU <RowMatrixXd> Lu(ori);
-	return ori.completeOrthogonalDecomposition().pseudoInverse();
-//    if (Lu.isInvertible())
-//        return ori.inverse();
-//    else
-//        return ori.completeOrthogonalDecomposition().pseudoInverse();
-
-
-}
-
-double standard_normalCDF(double x) // Phi(-∞, x) aka N(x)
-{
-    return std::erfc(-x / std::sqrt(2)) / 2;
-}
-
-/* The function checks if the array elements are consecutive
-If elements are consecutive, then returns true, else returns
-false */
-bool areConsecutive(int* arr, int n)  //int arr[]
-{
-	//int n = sizeof(arr) / sizeof(arr[0]);
-	if (n <  1)
-		return false;
-
-	int min = *min_element(arr, arr + n);
-
-	int max = *max_element(arr, arr + n);
-
-	if (max - min + 1 == n)
-	{
-		/* Create a temp array to hold visited flag of all elements.
-		Note that, calloc is used here so that all values are initialized
-		as false */
-		bool *visited = (bool *)calloc(n, sizeof(bool));
-		int i;
-		for (i = 0; i < n; i++)
-		{
-			/* If we see an element again, then return false */
-			if (visited[arr[i] - min] != false)
-				return false;
-
-			/* If visited first time, then mark the element as visited */
-			visited[arr[i] - min] = true;
-		}
-
-		/* If all elements occur once, then return true */
-		return true;
-	}
-
-	return false; // if (max - min  + 1 != n)
-}
-
- 
-//void tokenize(std::string const &str, const char* delim,
-//            std::vector<std::string> &out)
-//{
-//    char *token = strtok_s(const_cast<char*>(str.c_str()), delim);
-//    while (token != nullptr)
-//    {
-//        out.push_back(std::string(token));
-//        token = strtok_s(nullptr, delim);
-//    }
-//}
-
-vector<string> splitString(string str, char splitter) {
-	vector<string> result;
-	string current = "";
-	for (int i = 0; i < str.size(); i++) {
-		if (str[i] == splitter) {
-			if (current != "") {
-				result.push_back(current);
-				current = "";
-			}
-			continue;
-		}
-		current += str[i];
-	}
-	if (current.size() != 0)
-		result.push_back(current);
-	return result;
-}
-bool variable_exists(string var_name,  vector<string> &cols) {
-	//for (string var_name : names)
-	//	if (var_name == var_name)
-	//		return true;
-
-	//return false;
-
-
-	// if (std::find(cols.begin(), cols.end(), var_name) == cols.end())
-	// 	return false;
-	// return true;
[…]
