+++ pydynpd/sandbox/pydynpd/build_upload.sh
@@ -1,3 +1,0 @@
-rm -rf dist/*
-python3 -m build
-python3 -m twine upload --repository pypi dist/*.whl

[…]
