### 2022-05-24 #4416 @editorialbot
[REVIEW]: pydynpd: A python package for dynamic panel model
**Submitting author:** @dazhwu (<a href="http://orcid.org/0000-0003-0304-3332">Dazhong Wu</a>)
**Repository:** https://github.com/dazhwu/pydynpd
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.5
