### 2022-06-29 #4416 @dazhwu
Hi,
I just searched in github and from a repository (https://github.com/mikldk/jossrevsy/blob/main/global.R) I opened the following worksheet:

https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822

I searched among those records for reviewers who meet "regression" and "python", I got the following matches. Hope it helps.

username	Preferred Programming Languages	Other Programming Languages	email
----------------------------------------------------------------------------------------------------------
ejolly	Python	R, Matlab, Javascript	eshin.jolly.gr@dartmouth.edu
a-slide	Python	"BashCython"	aleg@ebi.ac.uk
RomainAzais	"RPython"		romain.azais@gmail.com
potterzot	R, Python, Stata		potter.nicholas@gmail.com
mmrabe	R	Python, C	maximilian.rabe@uni-potsdam.de
Athene-ai	R, Python	Julia	danielacialfi@gmail.com
lbluque	Python, Cython	C++, C, julia	lbluque@berkeley.edu
