### 2022-07-05 #4416 @chartgerink
Thanks @mmrabe! Enjoy your parental leave and I appreciate you letting us know.

@lschwetlick would you be interested in reviewing this software and paper for the Journal of Open Source Software? You can find the submitted manuscript [on GitHub](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04416/joss.04416/10.21105.joss.04416.pdf).

For a bit more information:
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged.

> JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html
