### 2022-07-08 #4416 @Athene-ai
> > > Thanks @dazhwu!
> > > For @a-slide @RomainAzais @potterzot @mmrabe @Athene-ai @lbluque - might one/some of you be interested to review this paper on dynamic panel modeling?
> > > Find [the paper on GitHub](https://github.com/openjournals/joss-papers/blob/joss.04416/joss.04416/10.21105.joss.04416.pdf), and [the source code](https://github.com/dazhwu/pydynpd) as well.
> > > I would greatly appreciate any help as we've had issues finding reviewers for this paper.
> >
> >
> > I am very delighted to review this paper
>
> Thank you so much!

The pleasure is mine !
