### 2022-07-11 #4416 @chartgerink
Thanks @Athene-ai for helping in this process - much appreciated 🙏

You can comment on this thread with `@editorialbot generate my checklist` and your checklist will be generated 😊
