### 2022-07-11 #4416 @RomainAzais
Hello @chartgerink
Thank you for considering me as a potential reviewer of this paper.
Unfortunately, I won't be available before mid-September for this kind of task.
Best regards

> For @a-slide @RomainAzais @potterzot @mmrabe @Athene-ai @lbluque - might one/some of you be interested to review this paper on dynamic panel modeling?
