### 2022-07-18 23102ba
bug fixed
removed a bug related to system GMM when  forward orthogonal deviation transformation is used
