+++ pydynpd/panel_data.py
@@ -3,7 +3,7 @@ import math
 import numpy as np
 from pandas import DataFrame
 
-from pydynpd.common_functions import get_first_diff_table
+from pydynpd.common_functions import get_first_diff_table, get_fod_table
 from pydynpd.info import options_info
 
 
@@ -34,7 +34,8 @@ class panel_data():
         num_cols = self.data.shape[1]
 
         self.fd_data = get_first_diff_table(self.data[:, range(0, num_cols)], self.N)
-        # self.fod_data=self.fod_trans(self.data, self.N, [1, num_cols])
+        if (options.transformation=='fod') & (options.level==False):
+            self.fod_data=get_fod_table(self.data, self.N)
 
     def xtset(self, df: DataFrame, _individual, _time):
         df.sort_values(by=[_individual, _time])
@@ -103,8 +104,7 @@ class panel_data():
         else:
             last=T-3
         start = last + 1 - height
-        print(np.around(temp,3))
-        print(start)
+
         D=temp[start:(last+1),:]
-        print(np.around(D,3))
+
         return (D)
