### 2022-07-20 d9ff233
Removed a bug on the calculation of number of observations for system GMM
