### 2022-07-29 #4416 @Athene-ai
Thanks @chartgerink for your suggestion.

Yes, _Functionality_  part should be better implemented especially in giving more details about the API method implementation
