+++ vignettes/API.md
@@ -1,0 +1,34 @@
+This document describes how to manipulate the outputs of the abond function. For instruction on what inputs are needed to run the abond function, please refer to
+
+## outputs of the abond function
+
+Estimates dynamic panel model using difference GMM or system GMM.
+
++-----------------------------------------------------------------------------------------------------+
+| ![Figure 1](https://raw.githubusercontent.com/dazhwu/pydynpd/main/vignettes/images/traditional.png) |
++:===================================================================================================:+
+| *Figure 1*                                                                                          |
++-----------------------------------------------------------------------------------------------------+
+
+### Usage:
+
+regression.abond()
+
+Quarto enables you to weave together content and executable code into a finished document. To learn more about Quarto see <https://quarto.org>.
+
+## Running Code
+
+When you click the **Render** button a document will be generated that includes both content and the output of embedded code. You can embed code like this:
+
+```{python}
+1 + 1
+```
+
+You can add options to executable code like this
+
+```{python}
+#| echo: false
+2 * 2
+```
+
+The `echo: false` option disables the printing of code (only output is displayed).
