+++ vignettes/API.md
@@ -1,19 +1,106 @@
 This document describes how to manipulate the outputs of the abond function. For instruction on what inputs are needed to run the abond function, please refer to [Tutorial](https://github.com/dazhwu/pydynpd/blob/main/vignettes/Tutorial.ipynb).
 
-## Outputs of the abond function
+## List of models returned
 
-After regression, the abond function returns a list of models:
+After regression, the abond function returns an "abond" object whose major property is "models" representing a list of models:
 
 ![Figure 1](https://raw.githubusercontent.com/dazhwu/pydynpd/main/vignettes/Images/list_models.svg)
 
-Multiple models could be returned because users may include "?" in command string. For example, in Example 8 in [Tutorial](https://github.com/dazhwu/pydynpd/blob/main/vignettes/Tutorial.ipynb), the following comand string generates 5 models:
+Multiple models could be returned because users may include "?" in command string. For example, in Example 8 in [Tutorial](https://github.com/dazhwu/pydynpd/blob/main/vignettes/Tutorial.ipynb), the following comand string generates 5 models: ··· command_str='n L(1:?).n w k \| gmm(n, 2:3) pred(w k)' mydpd = regression.abond(command_str, df, ['id', 'year']) ···
 
-    command_str='n L(1:?).n w k  | gmm(n, 2:3) pred(w k)'
-    mydpd = regression.abond(command_str, df, ['id', 'year'])
+What model(s) to be included in the list returned is based on the following rule:
 
-### Usage:
+-   If the user specifies THE model in the command string (i.e., she doesn't use "?" in command string), then this model is included in the list which contains only one model.
 
-regression.abond()
+-   Otherwise, only models that statisfy all of the following conditions are included in the list returned:
+
+    -   pass Hansen over-identification test (i.e., its P value should be greater than 5%)
+
+    -   pass AR(2) test (i.e., its P value is greater than 5%)
+
+    -   The P-value of Hansen over-identification test shouldn't be too high. In other words, it should be less than 99.99%. This is because a P value too close to 1 indicates a potential too-many-instrument issue.
+
+    For example, in Example 8 in [Tutorial](https://github.com/dazhwu/pydynpd/blob/main/vignettes/Tutorial.ipynb), 5 models are generated but only 3 of them are returned because the other two fail to satisfy at least one of the conditions above.
+
+    We can access each model on the list by its index which starts from 0. For example, the following code runs regression and then accesses the first model in the list returned:
+
+        command_str='n L(1:?).n w k \| gmm(n, 2:3) pred(w k)'
+        mydpd = regression.abond(command_str, df, ['id', 'year'])
+        m=mydpd.models[0]
+
+In the example above, mydpd is an "abond" object, and its "models" property represents a list of models. On the other hand, m, as a dynamic panel model, is the first model on the list. Next, we discuss the properties of the dynamic panel model object.
+
+## Dynamic penel model object
+
+The list below shows the properties of a dynamic panel model object:
+
+| Property         | Data Type                   | Meaning                                               |
+|------------------|:------------------|:----------------------------------|
+| AR_list          | list                        | list of Arellano-Bond test objects                    |
+| MMSC_LU          | dictionary                  | contains bic, aic, and hqic values                    |
+| N                | int                         | number of individuals                                 |
+| T                | int                         | number of time periods                                |
+| command_str      | str                         | command string of the model                           |
+| final_xy_tables  | dictionary                  | contains data of dependent and independent variables. |
+| hansen           | hansen test object          | hansen over identification test                       |
[…]
