### 2022-07-30 #4416 @Athene-ai
> Hi Daniela,
>
> Thank you for your comments. Right now, we have the input part of the API document in [Tutorial](https://github.com/dazhwu/pydynpd/blob/main/vignettes/Tutorial.ipynb). I am now adding the output part of it. After I have finished, I will let you know.
>
> Thanks,

Perfect ;-)
