+++ vignettes/API.md
@@ -37,24 +37,29 @@ In the example above, mydpd is an "abond" object, and its "models" property repr
 
 The list below shows the properties of a dynamic panel model object:
 
-| Property         | Data Type                   | Meaning                                               |
-|------------------|:----------------------------|:------------------------------------------------------|
-| AR_list          | list                        | list of Arellano-Bond test objects                    |
-| MMSC_LU          | dictionary                  | contains bic, aic, and hqic values                    |
-| N                | int                         | number of individuals                                 |
-| T                | int                         | number of time periods                                |
-| command_str      | str                         | command string of the model                           |
-| final_xy_tables  | dictionary                  | contains data of dependent and independent variables. |
-| hansen           | hansen test object          | hansen over identification test                       |
-| avg_obs          | int                         | average number of observations per individual         |
-| max_obs          | int                         | maximum number of observations per individual         |
-| min_obs          | int                         | minimum number of observations per individual         |
-| num_obs          | int                         | total number of observations                          |
-| regression_table | pandas data frame           | contains all statistics in the regression table       |
-| step_results     | list of step result objects | discussed                                             |
-| z_list           | Numpy ndarray               | instrument matrix                                     |
+| Property         | Data Type                   | Meaning                                                                                                                                                                                 |
+|------------------|:----------------|:----------------------------------|
+| AR_list          | list                        | list of Arellano-Bond test objects                                                                                                                                                      |
+| MMSC_LU          | dictionary                  | contains bic (Bayesian information criterion), aic (Akaike information criterion), and hqic (Hannan-Quinn information criterion) values suggested by Andrews, D. and Lu, B. (2001) (\*) |
+| N                | int                         | number of individuals                                                                                                                                                                   |
+| T                | int                         | number of time periods                                                                                                                                                                  |
+| command_str      | str                         | command string of the model                                                                                                                                                             |
+| final_xy_tables  | dictionary                  | contains data of dependent and independent variables.                                                                                                                                   |
+| hansen           | hansen test object          | hansen over identification test                                                                                                                                                         |
+| avg_obs          | int                         | average number of observations per individual                                                                                                                                           |
+| max_obs          | int                         | maximum number of observations per individual                                                                                                                                           |
+| min_obs          | int                         | minimum number of observations per individual                                                                                                                                           |
+| num_obs          | int                         | total number of observations                                                                                                                                                            |
+| regression_table | pandas data frame           | contains all statistics in the regression table                                                                                                                                         |
+| step_results     | list of step result objects | discussed                                                                                                                                                                               |
+| z_list           | Numpy ndarray               | instrument matrix                                                                                                                                                                       |
+
+(\*) Andrews, D. and Lu, B. (2001). Consistent Model and Moment Selection Procedures for
+GMM Estimation with application to dynamic panel data models. Journal of Econometrics, 101(1):123--164.
 
 ### AR list property
+
[…]
