+++ vignettes/API.md
@@ -38,13 +38,13 @@ In the example above, mydpd is an "abond" object, and its "models" property repr
 The list below shows the properties of a dynamic panel model object:
 
 | Property         | Data Type                   | Meaning                                                                                                                                                                                 |
-|------------------|:----------------|:----------------------------------|
+|------------------|:----------------------------|:----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | AR_list          | list                        | list of Arellano-Bond test objects                                                                                                                                                      |
 | MMSC_LU          | dictionary                  | contains bic (Bayesian information criterion), aic (Akaike information criterion), and hqic (Hannan-Quinn information criterion) values suggested by Andrews, D. and Lu, B. (2001) (\*) |
 | N                | int                         | number of individuals                                                                                                                                                                   |
 | T                | int                         | number of time periods                                                                                                                                                                  |
 | command_str      | str                         | command string of the model                                                                                                                                                             |
-| final_xy_tables  | dictionary                  | contains data of dependent and independent variables.                                                                                                                                   |
+| final_xy_tables  | dictionary                  | each element of the dictionary is a Numpy ndarray that contains data of dependent variable or independent variables.                                                                    |
 | hansen           | hansen test object          | hansen over identification test                                                                                                                                                         |
 | avg_obs          | int                         | average number of observations per individual                                                                                                                                           |
 | max_obs          | int                         | maximum number of observations per individual                                                                                                                                           |
@@ -54,8 +54,7 @@ The list below shows the properties of a dynamic panel model object:
 | step_results     | list of step result objects | discussed                                                                                                                                                                               |
 | z_list           | Numpy ndarray               | instrument matrix                                                                                                                                                                       |
 
-(\*) Andrews, D. and Lu, B. (2001). Consistent Model and Moment Selection Procedures for
-GMM Estimation with application to dynamic panel data models. Journal of Econometrics, 101(1):123--164.
+(\*) Andrews, D. and Lu, B. (2001). Consistent Model and Moment Selection Procedures for GMM Estimation with application to dynamic panel data models. Journal of Econometrics, 101(1):123--164.
 
 ### AR list property
 
@@ -68,7 +67,7 @@ This list contains two AR test (Arellano-Bond test) objects. For example, if m i
 An AR test object has three properties:
 
 | Property | Data Type | Meaning                                                               |
-|------------------|:-----------------|:-----------------------------------|
+|----------|:----------|:----------------------------------------------------------------------|
 | AR       | float     | AR test value                                                         |
 | P_value  | float     | P value of the AR test                                                |
 | lag      | int       | 1 for the first order test (AR(1)) and 2 for the second-order (AR(2)) |
@@ -132,7 +131,17 @@ The output is too long and therefore we only show below the section generated by
 
 ### final_xy_tables property
 
+This attribute is a dictionary. What are included in the dictionary depends on the transformation method used. For example, suppose m is a dynamic panel model object. If first difference method is used, the dictionary has two elements: m.final_xy_tables['Cx'] and m.final_xy_tables['Cy'] contain data of independent and dependent variables respectively. On the other hand, if forward orthogonal deviation transformation is used, then the dictionary has four elements: m.final_xy_tables['Cx'], m.final_xy_tables['Cy'], m.final_xy_tables['Diff_x'], and m.final_xy_tables['Diff_y']. The latter two elements are the first-difference matrices of independent and dependent variables respectively.
+
 ### hensen property
+
+It is an object that has the following attributes:
+
+| Property   | Data Type | Meaning                    |
+|------------|:----------|:---------------------------|
+| test_value | float     | Hensen test value          |
+| P_value    | float     | P value of the Hensen test |
+| df         | int       | degree of freedom          |
 
 ### regression_table property
 
