+++ vignettes/API.md
@@ -131,15 +131,15 @@ The output is too long and therefore we only show below the section generated by
 
 This attribute is a dictionary. What are included in the dictionary depends on the transformation method used. For example, suppose m is a dynamic panel model object. If first difference method is used, the dictionary has two elements: m.final_xy_tables['Cx'] and m.final_xy_tables['Cy'] contain data of independent and dependent variables respectively. On the other hand, if forward orthogonal deviation transformation is used, then the dictionary has four elements: m.final_xy_tables['Cx'], m.final_xy_tables['Cy'], m.final_xy_tables['Diff_x'], and m.final_xy_tables['Diff_y']. The latter two elements are the first-difference matrices of independent and dependent variables respectively.
 
-### hensen property
+### hansen property
 
 It is an object that has the following attributes:
 
 | Property   | Data Type | Meaning                    | Example (suppose m is a dynamic panel model object)   |
 |------------|:----------|:---------------------------|:--------------|
-| test_value | float     | Hensen test value          |m.hensen.test_value |
-| p_value    | float     | P value of the Hensen test |m.hensen.p_value |
-| df         | int       | degree of freedom          |m.hensen.df |
+| test_value | float     | Hansen test value          |m.hansen.test_value |
+| p_value    | float     | P value of the Hansen test |m.hansen.p_value |
+| df         | int       | degree of freedom          |m.hansen.df |
 
 ### regression_table property
 
