### 2022-07-31 #4416 @Athene-ai
> Hi Daniela,
>
> I have added a new document to the github web site. Now, there are two documents regarding how to use the software.
>
> Document 1 is on what input parameters are needed to run the command and how to form them. The link is [here](https://github.com/dazhwu/pydynpd/blob/main/vignettes/Tutorial.ipynb).
>
> Document 2 is the new one I just added. It discusses the outputs (i.e., the objects and their attributes generated) and how to manipulate them. The link is [here](https://github.com/dazhwu/pydynpd/blob/main/vignettes/API.md).
>
> Each of the two documents mentions the other at the beginning of it. I also include the two links in the home page (README.md).
>
> Please let me know if there is anything I didn't include. Thanks,

I think that these two documents are fine
