### 2022-08-09 #4416 @chartgerink
Okay, thanks @Athene-ai.

Did your discussion with @dazhwu help resolve the open item on the checklist though? Namely:

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
