### 2022-12-07 #4416 @dazhwu
I found a new match on google sheet:

mhu48	R, Python, C++, SQL, SAS	NONMEM	University of California Santa Barbara	mingzhaohu@ucsb.edu
