### 2023-02-16 #4416 @chartgerink
@dazhwu thanks for your incredible patience with this process 🙌

This paper is accepted for publication in JOSS, pending a few minor revisions. 🥳

1. "no python nor Julia" in line 36 - this maybe should be "no python or Julia"?
2. Could you create a larger version of Figure 1? It is hard to read at this point
3. line 55 - maybe it should read "panel models" (plural)?
4. Could you add figure captions to all figures? (this is a must)
5. You could consider joining figures 2 and 3 in one, with two panels. This may help legibility of the differences (only a suggestion that you get to decide on 😊 )
6. It may be an idea to add a slight conclusion simply to help the reader finish the piece. It now ends on an equation, which is unlikely to be remembered. What's your take home message in two or three sentences? (I leave this up to you)

After these final revisions are made I will do a quick check. If that all works out, I will request you to tag and archive the repository (e.g., Zenodo, Figshare) and share the DOI here.

Thanks for this important paper 👍
