+++ vignettes/paper.md
@@ -61,11 +61,8 @@ packages which rely heavily on R-specific components (that is a main
 reason they are not fast), our code uses components common to any
 programming language, making it easy to translate to R or Julia.
 
-<p align="center">
-  <img alt="img-name" src="Test_1.svg" width="1000">
-  <br>
-    <em>Figure 1: running time (relative to the fastest)</em>
-</p>
+![Figure 1: running time (relative to the fastest).\label{fig:example}](Test+1.svg)
+
 
 
 # The pydynpd package 
