### 2023-03-04 #4416 @dazhwu
> t

Hi,

Below is the list of modifications I made:

1) "no python nor Julia"  has been changed to "no python or Julia".
2) Figure 1 is much larger and more readable.
3) I changed "panel model" on line 55 to "panel models".
4) All of the three figures have captions now.
5) I put figure 2 and figure 3 next to each other vertically. I tried to put them side by side, but then each would be small.
6) I moved the equation originally at the end to above the last paragraph. Add add the following sentence above the equation:

In other words, users may use pydynpd to estimate the following model with question markers indicating values not determined yet:

Please let me know if there is any change I need to make. Thanks,
