### 2023-03-06 #4416 @dazhwu
> @dazhwu thanks! We need the title and authors on the archived repository to match the paper. Could you please edit this for us?

Is this the one you request? It is included on https://joss.readthedocs.io/en/latest/submitting.html

**create a metadata file describing your software and include it in your repository. We provide [a script](https://gist.github.com/arfon/478b2ed49e11f984d6fb) that automates the generation of this metadata.**
