+++ README.md
@@ -1,4 +1,4 @@
-# pydynpd: Dynamic panel estimation for Difference and System GMM (generalized method-of-moments)
+# pydynpd: A python package for dynamic panel model
 [![DOI](https://zenodo.org/badge/466146436.svg)](https://zenodo.org/badge/latestdoi/466146436)
 [![pypi package](https://img.shields.io/pypi/v/pydynpd?style=plastic)](https://pypi.org/project/pydynpd/)
 
