### 2023-03-07 #4416 @arfon
@dazhwu – I made a number of small edits to your paper here: https://github.com/dazhwu/pydynpd/pull/6
